board/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.4
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
easy-thumbnails==2.7.1
Pillow==8.2.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
testfixtures==6.17.1
